+++ src/main/java/roomescape/dao/ReservationDao.java
@@ -7,9 +7,9 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationCommand;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.Reservation.Reservation;
+import roomescape.domain.Reservation.ReservationCommand;
+import roomescape.domain.ReservationTime.ReservationTime;
 
 @Repository
 public class ReservationDao {
