--- src/main/java/roomescape/domain/Reservation.java
@@ -1,4 +1,0 @@
-package roomescape.domain;
-
-public record Reservation(long id, String name, String date, ReservationTime time) {
-}
