+++ src/main/java/roomescape/domain/Reservation/Reservation.java
@@ -1,0 +1,6 @@
+package roomescape.domain.Reservation;
+
+import roomescape.domain.ReservationTime.ReservationTime;
+
+public record Reservation(long id, String name, String date, ReservationTime time) {
+}
