+++ src/main/java/roomescape/domain/Reservation/ReservationCommand.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.Reservation;
 
 public record ReservationCommand(String name, String date,  long timeId) {
 }
