+++ src/main/java/roomescape/domain/ReservationTime/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.ReservationTime;
 
 public record ReservationTime(long id, String startAt) {
 }
