+++ src/main/java/roomescape/domain/ReservationTime/ReservationTimeCommand.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.ReservationTime;
 
 public record ReservationTimeCommand(String startAt) {
 }
