+++ src/main/java/roomescape/dto/ReservationTime/AddReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.ReservationTime;
 
 public record AddReservationTimeRequest(String startAt) {
 }
