+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Repository;
 import roomescape.dao.ReservationTimeDao;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.ReservationTimeCommand;
+import roomescape.domain.ReservationTime.ReservationTime;
+import roomescape.domain.ReservationTime.ReservationTimeCommand;
 
 @Repository
 public class ReservationTimeRepository {
